+++ lib/page-encointer/ceremony/attestation/qrCode.dart
