+++ lib/page-encointer/ceremony/attestation/scanQrCode.dart
