+++ lib/page-encointer/ceremony/confirmAttendeesDialog.dart
