+++ lib/page-encointer/phases/attesting/attestingPage.dart
@@ -5,8 +5,8 @@ import 'package:mobx/mobx.dart';
 import 'package:polka_wallet/common/components/passwordInputDialog.dart';
 import 'package:polka_wallet/common/components/roundedButton.dart';
 import 'package:polka_wallet/common/components/roundedCard.dart';
-import 'package:polka_wallet/page-encointer/attesting/confirmAttendeesDialog.dart';
-import 'package:polka_wallet/page-encointer/attesting/meetupPage.dart';
+import 'package:polka_wallet/page-encointer/ceremony/ceremonyPage.dart';
+import 'package:polka_wallet/page-encointer/ceremony/confirmAttendeesDialog.dart';
 import 'package:polka_wallet/page-encointer/common/assignmentPanel.dart';
 import 'package:polka_wallet/page/account/txConfirmPage.dart';
 import 'package:polka_wallet/service/substrateApi/api.dart';
