+++ lib/page-encointer/phases/registering/registerParticipantPanel.dart
